move polyfill tests to own file
